ml-experiments/requirements.txt
torch>=2.9.0,<3
torchvision>=0.24.0,<1
numpy>=2.1.0
Pillow>=10.4.0
scikit-learn>=1.5.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.9
httpx>=0.27.0
python-dotenv>=1.0.0
matplotlib>=3.9.0
